Adding tutorial 14 - my terrain has reflectivity and damper variables, both in the java and shader code.
That's because my tutorials are sequentially built (tutorial 14 adds to 13). It appears that the code in the youtube videos is not always sequential because these variables are missing from thin matrix's presentation. This doesn't make a big difference in the appearance of the terrain. In fact, I don't think it makes any difference at this point. You can try to remove the variables from the shader as an exercise.